Revoking an object URL immediately after triggering navigation causes navigation to fail https://bugs.webkit.org/show_bug.cgi?id=212279 <rdar://problem/63553090>
Reviewed by Geoffrey Garen.

Source/WebCore:

When doing a policy check for a Blob URL, we clone the blob and create a new temporary Blob URL
that stays alive for the duration of the policy check. We made sure to update the ResourceRequest
URL with the new Blob URL, however, we were failing to update the DocumentLoader's request.
As a result, if the client responded with Policy USE, the DocumentLoader would still attempt to
navigate to the old Blob URL.

Test: fast/loader/revoke-blob-url-after-navigation.html

* loader/PolicyChecker.cpp:
(WebCore::FrameLoader::PolicyChecker::extendBlobURLLifetimeIfNecessary const):
(WebCore::FrameLoader::PolicyChecker::checkNavigationPolicy):
(WebCore::FrameLoader::PolicyChecker::checkNewWindowPolicy):
* loader/PolicyChecker.h:

LayoutTests:

Add layout test coverage.

* fast/loader/revoke-blob-url-after-navigation-expected.txt: Added.
* fast/loader/revoke-blob-url-after-navigation.html: Added.


Identifier: 225148@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262085 268f45cc-cd09-0410-ab3c-d52691b4dbfc