[ macOS iOS ] REGRESSION(r261600?): imported/w3c/web-platform-tests/html/dom/reflection-embedded.html & imported/w3c/web-platform-tests/html/dom/reflection-forms.html are flaky failures https://bugs.webkit.org/show_bug.cgi?id=212430
Reviewed by Saam Barati.

JSTests:

* stress/is-constructor-should-be-correct-in-ai.js: Added.
(shouldBe):
(test):

Source/JavaScriptCore:

r261600 added IsConstructor rule to DFG AI. That rule is saying that if the object is not JSFunction and not ProxyObject,
then it must not be a constructor. But this is wrong since any objects can implement getConstructData and DOM constructors
are actually implementing it while it is not JSFunction and it is not ProxyObject. This patch removes that rule.

* dfg/DFGAbstractInterpreterInlines.h:
(JSC::DFG::AbstractInterpreter<AbstractStateType>::executeEffects):
* tools/JSDollarVM.cpp:


Identifier: 225290@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262231 268f45cc-cd09-0410-ab3c-d52691b4dbfc