UIColor and NSColor WebCore::Color factories should return invalid colors for nil input colors https://bugs.webkit.org/show_bug.cgi?id=212631
Reviewed by Anders Carlsson.

* platform/graphics/mac/ColorMac.mm:
(WebCore::colorFromNSColor):
(WebCore::semanticColorFromNSColor):
* platform/ios/ColorIOS.mm:
(WebCore::colorFromUIColor):
This doesn't affect any code currently in WebKit, but it is very, very surprising
that these functions happily accept a null color, assert in debug, but in release
do crazy things like try to paint the null color into a small bitmap to figure out
what it really is.

Also, this matches the behavior of the Color constructors that take CGColorRef.


Identifier: 225475@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262428 268f45cc-cd09-0410-ab3c-d52691b4dbfc