[EME][GStreamer] cdmProxyAttached does not need to force a bump ref in the signature https://bugs.webkit.org/show_bug.cgi?id=212754
Reviewed by Philippe Normand.

cdmProxyAttached is currently receiving a RefPtr<CDMProxy> in the
signature, what causes a ref bump when the function is called. A
const RefPtr<CDMProxy>& is more suitable cause the reference is
already bumped when the CDMProxy assigned in the decryptor
attribute.

No new tests, just a rework.

* platform/graphics/gstreamer/eme/WebKitClearKeyDecryptorGStreamer.cpp:
(cdmProxyAttached):
* platform/graphics/gstreamer/eme/WebKitCommonEncryptionDecryptorGStreamer.h:


Identifier: 225565@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262545 268f45cc-cd09-0410-ab3c-d52691b4dbfc