Reduce telemetry and report generation for Network Process sandbox https://bugs.webkit.org/show_bug.cgi?id=212778 <rdar://problem/63891745>
Reviewed by Per Arne Vollan.

This patch cleans up a set of telemetry and reports generated during normal
WebKit operation. Now that we understand the uses of these XPC services, we
can reduce the logging traffic.

This is important to do now, since telemetry and report generation has
battery life and performance impact.

* NetworkProcess/mac/com.apple.WebKit.NetworkProcess.sb.in:
* Resources/SandboxProfiles/ios/com.apple.WebKit.Networking.sb:


Identifier: 225585@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262572 268f45cc-cd09-0410-ab3c-d52691b4dbfc