Reduce telemetry and report generation for WebContent Process sandbox https://bugs.webkit.org/show_bug.cgi?id=212779 <rdar://problem/63769879>
Reviewed by Per Arne Vollan.

This patch cleans up a set of telemetry and reports generated during normal
WebKit operation. Now that we understand the uses of these XPC services, we
can reduce the logging traffic.

This is important to do now, since generating the telemetry and reports has
a non-zero power cost.

* Resources/SandboxProfiles/ios/com.apple.WebKit.WebContent.sb:
* WebProcess/com.apple.WebProcess.sb.in:


Identifier: 225592@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262586 268f45cc-cd09-0410-ab3c-d52691b4dbfc