Nullptr crash in DeleteSelectionCommand::doApply() when ending position is disconnected. https://bugs.webkit.org/show_bug.cgi?id=212723 <rdar://problem/63866653>
Reviewed by Geoffrey Garen.

Source/WebCore:

In this test case, while merging paragraphs after deleting a text element, we need call removeNodeAndPruneAncestors()
to remove a BR node. However, the ancestor of BR is also removed. Later we try to insert a node at the parent of the
removed ancestor in function DeleteSelectionCommand::doApply().

For now we just check the parentless inserting position and bail out. The proper fix should be re-designing
removeNodeAndPruneAncestors() or select a different inserting position after removeNodeAndPruneAncestors() is called.

Test: editing/deleting/delete-txt-in-dl-crash.html

* editing/DeleteSelectionCommand.cpp:
(WebCore::DeleteSelectionCommand::doApply):

LayoutTests:

Added a regression test for the crash.

* editing/deleting/delete-txt-in-dl-crash-expected.txt: Added.
* editing/deleting/delete-txt-in-dl-crash.html: Added.


Identifier: 225599@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262593 268f45cc-cd09-0410-ab3c-d52691b4dbfc