Make CodeBlockHash robust against unreasonably long source code. https://bugs.webkit.org/show_bug.cgi?id=212847 <rdar://problem/64024279>
Reviewed by Saam Barati.

This patch adds a heuristic to avoid trying to compute the CodeBlockHash on
unreasonably long source code strings.  This is done by first applying a length
check and, if needed, computing the hash with an alternate method.

This is OK to do because:
1. CodeBlockHash is not a critical hash.
2. In practice, reasonable source code are not that long.
3. And if they are that long, then we are still diversifying the hash on their
   length. But if they do collide, it's OK.

The only invariant here is that we should always produce the same hash for the
same source string.  Since the algorithm is deterministic, this invariant is not
violated.

* bytecode/CodeBlockHash.cpp:
(JSC::CodeBlockHash::CodeBlockHash):


Identifier: 225676@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262677 268f45cc-cd09-0410-ab3c-d52691b4dbfc