[JSC] Checksum for generated files should be emitted at the end of the files https://bugs.webkit.org/show_bug.cgi?id=212875
Reviewed by Mark Lam.

If the offlineasm file generation is interrupted in the middle of the generation, it already emitted checksum.
So next file generation can accept this broken file as a result of offlineasm and skip file generation.
We should emit checksum at the end of files. For now, this patch takes a quick way: just iterating lines, getting
a last line and use it for checksum comparison.

* generator/GeneratedFile.rb:
* offlineasm/asm.rb:


Identifier: 225691@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262694 268f45cc-cd09-0410-ab3c-d52691b4dbfc