BaseAudioSharedUnit should unmute its clients in case of suspension even if not having any audio unit https://bugs.webkit.org/show_bug.cgi?id=212970
Reviewed by Eric Carlson.

Source/WebCore:

CoreAudioCaptureSource(s), when muted, are now calling stopProducingData.
This will, in turn, make the BaseAudioSharedUnit stop and no longer have any audio unit.
In that case, when resume is called on the BaseAudioSharedUnit, it will exit early as the audio unit is null.
This will prevent to unmute the CoreAudioCaptureSource(s).

Fix this by removing the audio unit check in BaseAudioSharedUnit::resume.
Add infrastructure testing to be able to write a test.

Covered by added test.

* platform/mediastream/RealtimeMediaSource.h:
* platform/mediastream/mac/BaseAudioSharedUnit.cpp:
(WebCore::BaseAudioSharedUnit::resume):
* platform/mediastream/mac/CoreAudioCaptureSource.cpp:
(WebCore::CoreAudioCaptureSource::setInterruptedForTesting):
* platform/mediastream/mac/CoreAudioCaptureSource.h:
* testing/Internals.cpp:
(WebCore::Internals::isMediaStreamSourceInterrupted const):
* testing/Internals.h:
* testing/Internals.idl:

Tools:

Add an option to capture in UI process so that the same test can run on both MacOS and iOS.
This eases the testing infra to be able to get audio capture unit state in WebProcess as is the case in iOS.

* WebKitTestRunner/TestController.cpp:
(WTR::TestController::resetPreferencesToConsistentValues):
(WTR::updateTestOptionsFromTestHeader):
* WebKitTestRunner/TestOptions.h:
(WTR::TestOptions::hasSameInitializationOptions const):

LayoutTests:

* fast/mediastream/media-stream-track-interrupted-expected.txt:
* fast/mediastream/media-stream-track-interrupted.html:


Identifier: 225775@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@262798 268f45cc-cd09-0410-ab3c-d52691b4dbfc