Consistently use WTF_ATTRIBUTE_PRINTF in Assertions.[cpp,h] https://bugs.webkit.org/show_bug.cgi?id=213204
Patch by Michael Catanzaro <mcatanzaro@gnome.org> on 2020-06-15
Reviewed by Darin Adler.

Use WTF_ATTRIBUTE_PRINTF for all functions taking a printf format specifier. The first
argument should be the position of the format string itself (starting from 1), and the
second argument should be the position of the ... parameter pack, or 0 for functions that
take a va_list.

* wtf/Assertions.cpp:
* wtf/Assertions.h:

Identifier: 226000@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@263062 268f45cc-cd09-0410-ab3c-d52691b4dbfc